Minor optimization

+++ src/foundation/src/PDFsharp/src/PdfSharp/Fonts.OpenType/OpenTypeFontTables.cs
@@ -126,7 +126,7 @@ namespace PdfSharp.Fonts.OpenType
                     break;
                 for (var c = start; c <= end; c++)
                 {
-                    glyphList.Add((int)c);
+                    glyphList.Add(c);
                 }
             }
             return glyphList;
